requirements.txt
psycopg[binary]>=3.2.2
python-dotenv>=1.0.1
requests>=2.32.3
PyYAML>=6.0.2

dnspython>=2.6.1
tldextract>=5.1.2

beautifulsoup4>=4.12.3
lxml>=5.2.2

# qualità vita / logging (opzionali ma utili)
rich>=13.7.1
